Remove filter locator.
