+++ src/Manager.php
@@ -7,8 +7,6 @@ use Cake\Utility\Inflector;
 use InvalidArgumentException;
 use Search\Model\Filter\FilterCollection;
 use Search\Model\Filter\FilterCollectionInterface;
-use Search\Model\Filter\FilterLocator;
-use Search\Model\Filter\FilterLocatorInterface;
 
 /**
  * Search Manager Service Class
@@ -38,13 +36,6 @@ class Manager
     protected $_collection = 'default';
 
     /**
-     * Filter Locator
-     *
-     * @var \Search\Model\Filter\FilterLocatorInterface
-     */
-    protected $_filterLocator;
-
-    /**
      * Constructor
      *
      * @param \Cake\Datasource\RepositoryInterface $repository Repository
@@ -52,21 +43,7 @@ class Manager
     public function __construct(RepositoryInterface $repository)
     {
         $this->_repository = $repository;
-        $this->_filterLocator = new FilterLocator($this);
-        $this->_collections['default'] = new FilterCollection($this->_filterLocator);
-    }
-
-    /**
-     * Sets the filter locator
-     *
-     * @param \Search\Model\Filter\FilterLocatorInterface $filterLocator Filter Locator
-     * @return $this
-     */
-    public function setFilterLocator(FilterLocatorInterface $filterLocator)
-    {
-        $this->_filterLocator = $filterLocator;
-
-        return $this;
+        $this->_collections['default'] = new FilterCollection($this);
     }
 
     /**
@@ -118,7 +95,7 @@ class Manager
             ));
         }
 
-        return new $className($this->_filterLocator);
+        return new $className($this);
     }
 
     /**
@@ -130,7 +107,7 @@ class Manager
     public function useCollection($name)
     {
         if (!isset($this->_collections[$name])) {
-            $this->_collections[$name] = new FilterCollection($this->_filterLocator);
+            $this->_collections[$name] = new FilterCollection($this);
         }
         $this->_collection = $name;
